Refactor/test mocks (#139)
* Move mocks to separate files
