+++ video_xblock/tests/mocks/__init__.py
@@ -1,0 +1,3 @@
+"""
+Mocks for Video Xblock testing.
+"""
