+++ video_xblock/tests/mocks/vimeo.py
@@ -1,0 +1,29 @@
+"""
+Vimeo backend mocks.
+"""
+
+from video_xblock.tests.mocks.base import BaseMock
+
+
+class VimeoAuthMock(BaseMock):
+    """
+    Vimeo auth mock class.
+    """
+
+    pass
+
+
+class VimeoDefaultTranscriptsMock(BaseMock):
+    """
+    Temporary mock for vimeo player. Need to be updated.
+    """
+
+    pass
+
+
+class VimeoDownloadTranscriptMock(BaseMock):
+    """
+    Vimeo download transcript mock class.
+    """
+
+    pass
