+++ video_xblock/tests/test_video_xblock.py
@@ -12,7 +12,8 @@ from django.conf import settings
 from video_xblock import VideoXBlock
 from video_xblock.utils import ugettext as _
 from video_xblock.tests.base import VideoXBlockTestBase
-from video_xblock.tests.mocks import MockCourse
+from video_xblock.tests.mocks.base import MockCourse
+
 from video_xblock.constants import PlayerName
